Moved unit tests up a folder level from restruct_poc
