+++ tests/unit_tests/model/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Specify the source files for the unit test executable
+target_sources(
+    openturbine_unit_tests
+    PRIVATE
+    # Model-related tests
+    test_model.cpp
+)
