+++ tests/unit_tests/regression/test_rotating_beam.cpp
