+++ tests/unit_tests/solver/test_compute_number_of_non_zeros.cpp
