+++ tests/unit_tests/solver/test_populate_sparse_indices.cpp
