+++ tests/unit_tests/system/test_calculate_gyroscopic_matrix.cpp
