+++ tests/unit_tests/system/test_calculate_inertia_stiffness_matrix.cpp
