+++ tests/unit_tests/system/test_calculate_strain.cpp
