+++ tests/unit_tests/system/test_integrate_residual_vector.cpp
@@ -6,7 +6,6 @@
 
 #include "src/beams/beams.hpp"
 #include "src/system/integrate_residual_vector.hpp"
-#include "tests/unit_tests/restruct_poc/test_utilities.hpp"
 
 namespace openturbine::tests {
