+++ tests/unit_tests/system/test_update_node_state.cpp
